pyproject.toml
[project]
name = "mcp-server-learning"
version = "0.1.0"
description = "Agentic AI that informs you about the weather in a given area by query."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.59.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.1",
    "openai==0.28",
    "python-dotenv>=1.1.1",
]
